refactor(verify): split api.py into submodules (#380)
verification/api.py (~90K, over the Council review cap) is split into constants.py, schemas.py (incl. the validation regexes used by its validators), evidence_render.py, and file_ops.py — all VERBATIM moves, no logic changes. api.py re-exports every moved name for backward compatibility (18 test files import them) and drops to ~39K, back under the review cap. Test patch targets whose CALLEES moved (fetch/git internals now called within file_ops) are mechanically retargeted to file_ops; api-level patch targets (_build_verification_prompt callers) unchanged. Full suite green (2964 — identical count to pre-refactor); ruff clean.

Closes #380.
